+++ odoo/local-src/connector_esb/__manifest__.py
@@ -29,6 +29,7 @@
         'product_additional',
         'queue_job_cron',
         'sale_confirm_background',
+        'l10n_be_apb_tax',
     ],
     'website': 'http://www.camptocamp.com',
     'data': [